enable worker metrics back
