+++ internal/common/repo.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 
-	"github.com/opentracing/opentracing-go"
 	"github.com/razorpay/metro/pkg/logger"
 	"github.com/razorpay/metro/pkg/registry"
 )
